changed file structure to add community support
