--- frontend/src/app/CallPoliticians.module.css
@@ -1,116 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@400;500;600&display=swap");
-
-/* Outer container: match the width of other components */
-.callPoliticiansContainer {
-  width: 100%; /* Use full width */
-  max-width: 1441px; /* Match the max width of other components */
-  margin: 0 auto; /* Centers the container */
-  background-color: #fff; /* White background */
-  padding: 40px 20px; /* Adjust padding as needed */
-  box-sizing: border-box; /* Includes padding in total width, auto shrinks */
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-/* Title and subtitle styling */
-.title {
-  font-family: "Montserrat", sans-serif;
-  color: #ee6c4d;
-  font-size: 2.5rem;
-  text-align: center;
-  margin-bottom: 1rem;
-}
-
-.subtitle {
-  font-family: "Montserrat", sans-serif;
-  color: #2d5177;
-  text-align: center;
-  font-size: 1.2rem;
-  line-height: 1.5;
-  max-width: 800px; /* Adjust as needed */
-  margin: 0 auto 2rem auto;
-}
-
-/* Flexbox for four circles in one row */
-.politicianRow {
-  display: flex;
-  justify-content: space-around; /* Distribute circles evenly */
-  width: 100%;
-  max-width: 1000px; /* Adjust as needed */
-}
-
-.politicianCard {
-  display: flex;
-  flex-direction: column;
-  align-items: center; /* Center the content */
-}
-
-.circle {
-  width: 173px;
-  height: 173px;
-  background-color: #dedede;
-  border-radius: 50%;
-}
-
-.politicianInfo {
-  text-align: center; /* Center the text */
-  margin-top: 8px;
-}
-
-.politicianName {
-  font-family: "Montserrat", sans-serif;
-  color: #ee6c4d;
-  font-weight: 600;
-  margin-bottom: 4px;
-}
-
-.politicianDetails {
-  font-family: "Montserrat", sans-serif;
-  color: #2d5177;
-  font-size: 0.9rem;
-  white-space: nowrap;
-}
-
-/* MEDIA QUERIES */
-@media (max-width: 1439px) {
-  /* Adjusted to trigger below 1440px */
-  .callPoliticiansContainer {
-    width: 100%;
-    padding: 40px 20px;
-  }
-
-  .politicianGrid {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 32px;
-    max-width: 600px;
-    margin: 0 auto;
-  }
-
-  .circle {
-    width: 120px;
-    height: 120px;
-  }
-}
-
-@media (max-width: 768px) {
-  .title {
-    font-size: 2rem;
-  }
-
-  .subtitle {
-    font-size: 1rem;
-    max-width: 100%;
-  }
-
-  .politicianGrid {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 24px;
-    max-width: 300px;
-  }
-
-  .circle {
-    width: 100px;
-    height: 100px;
-  }
-}
