--- frontend/src/app/SignPetition.module.css
@@ -1,98 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@400;500;600&display=swap");
-
-.signPetitionGrid {
-  display: flex;
-  gap: 100px;
-  flex-direction: row;
-  width: 100%;
-  max-width: 1440px;
-  max-height: 635px;
-  height: auto;
-  margin: 0 auto;
-  background-color: #ffffff;
-  padding: 70px 160px;
-  justify-content: space-between;
-}
-
-/* Responsive adjustments */
-@media (max-width: 720px) {
-  .signPetitionGrid {
-    flex-direction: row; /* Stacks the content vertically when space is limited */
-  }
-}
-
-.signPetitionText {
-  display: flex;
-  width: 500.1;
-  height: 252.35px;
-  justify-content: center;
-  color: white;
-  flex-direction: column;
-  gap: 40px;
-}
-
-@font-face {
-  font-family: "Open Sans";
-  src: url("fonts/OpenSans-Regular.ttf");
-}
-
-.signPetitionText h2 {
-  color: #002c5f;
-  font-family: "Montserrat", sans-serif;
-  font-size: 40px;
-  font-style: normal;
-  font-weight: 700;
-  letter-spacing: 0.96px;
-  height: 48.76px;
-  letter-spacing: 0%;
-}
-
-.signPetitionText p {
-  color: black;
-  text-align: left;
-  font-family: "Open Sans";
-  font-style: normal;
-  font-weight: 400;
-  width: 500.1px;
-  height: 135px; /* 133.333% */
-  font-size: 20px;
-  align-self: center;
-  line-height: 27.24px;
-  letter-spacing: 0%;
-}
-
-.signPetitionBottom {
-  width: 514.56px;
-  height: 67.66px;
-  display: flex;
-  flex-direction: row;
-  gap: 19px;
-  align-items: center;
-  white-space: nowrap;
-}
-
-.signPetitionBottomLeft {
-  display: flex;
-  flex-direction: column;
-  width: 429.46px;
-  height: 66px;
-  justify-content: center;
-}
-
-.signPetitionBottomLeft p {
-  text-align: center;
-  font-size: 24px;
-  font-weight: 400;
-  font-family: "Open Sans";
-  letter-spacing: 0%;
-}
-.signPetitionBottomLeft a {
-  font-size: 23px;
-  font-weight: 400;
-  font-family: "Open Sans";
-  letter-spacing: 0%;
-  color: black;
-  text-align: start;
-  text-decoration: underline;
-  font-style: italic;
-}
