+++ frontend/src/components/community_support/CommunitySupport.tsx
@@ -1,0 +1,13 @@
+import HelpOutToday from "@/components/community_support/HelpOutToday";
+import SampleScript from "@/components/community_support/SampleScript";
+import SignPetition from "@/components/community_support/SignPetition";
+
+export default function Solution() {
+  return (
+    <div className="">
+      <HelpOutToday />
+      <SampleScript />
+      <SignPetition />
+    </div>
+  );
+}
